+++ app/src/main/java/de/danoeh/antennapod/ui/episodeslist/EpisodeMultiSelectActionHandler.java
@@ -1,5 +1,6 @@
 package de.danoeh.antennapod.ui.episodeslist;
 
+import android.app.Activity;
 import android.util.Log;
 
 import androidx.annotation.PluralsRes;
@@ -9,21 +10,22 @@ import com.google.android.material.snackbar.Snackbar;
 import java.util.List;
 
 import de.danoeh.antennapod.R;
-import de.danoeh.antennapod.activity.MainActivity;
+import de.danoeh.antennapod.event.MessageEvent;
 import de.danoeh.antennapod.net.download.serviceinterface.DownloadServiceInterface;
 import de.danoeh.antennapod.storage.database.DBWriter;
 import de.danoeh.antennapod.storage.database.LongList;
 import de.danoeh.antennapod.model.feed.FeedItem;
 import de.danoeh.antennapod.ui.view.LocalDeleteModal;
+import org.greenrobot.eventbus.EventBus;
 
 public class EpisodeMultiSelectActionHandler {
     private static final String TAG = "EpisodeSelectHandler";
-    private final MainActivity activity;
+    private final Activity activity;
     private final int actionId;
     private int totalNumItems = 0;
     private Snackbar snackbar = null;
 
-    public EpisodeMultiSelectActionHandler(MainActivity activity, int actionId) {
+    public EpisodeMultiSelectActionHandler(Activity activity, int actionId) {
         this.activity = activity;
         this.actionId = actionId;
     }
@@ -120,7 +122,7 @@ public class EpisodeMultiSelectActionHandler {
                 snackbar.setText(text);
                 snackbar.show(); // Resets the timeout
             } else {
-                snackbar = activity.showSnackbarAbovePlayer(text, Snackbar.LENGTH_LONG);
+                EventBus.getDefault().post(new MessageEvent(text));
             }
         });
     }
